refactor: extract pipeline/install.rs
Move the core install pipeline (install_from_local_path,
install_bundle_file, install_items_of_kind,
install_with_name_resolution_from_local,
install_plugins_from_bundles) plus audience targeting helpers
(targets, audience_matches) out of pipeline/mod.rs into a focused
submodule. The audience tests move with their functions.

install_with_lockfile stays in mod.rs as the top-level coordinator
that ties install, hash, output, ancillary, and the lockfile
together.

Public API is preserved via 'pub use install::install_from_local_path'
in mod.rs. DoctorReport tests also relocate from mod.rs into report.rs
where the type lives.

Refs #180
